Refactor page button into browser component
