+++ zapzap/views/browser/__init__.py
@@ -1,3 +1,4 @@
+from .browser_page_button import BrowserPageButton
 from .browser_pages import BrowserPages
 from .browser_sidebar import BrowserSidebar, BrowserSidebarButton
 from .browser_view import BrowserView
